requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2.2
django-cors-headers==4.7.0
django-elasticsearch==7.0.2
django-elasticsearch-dsl==8.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
elastic-transport==8.17.1
elasticsearch==8.18.1
elasticsearch-dsl==8.18.0
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
multidict==6.5.0
oauthlib==3.2.2
packaging==25.0
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.20
yarl==1.20.1
